+++ tests/testthat/test-textmodel.R
@@ -51,7 +51,7 @@ test_that("textmodel_lss has all the attributes", {
     expect_identical(names(lss_test$beta), names(lss_test$frequency))
     expect_true(is.dfm(lss_test$data))
     expect_identical(lss_test$terms, feat_test)
-    expect_identical(names(lss_test$seeds_weighted), names(seedwords("pos-neg")))
+    expect_identical(names(lss_test$seeds), names(seedwords("pos-neg")))
 
     expect_equal(
         names(lss_test_fcm),
@@ -65,7 +65,7 @@ test_that("textmodel_lss has all the attributes", {
     expect_identical(names(lss_test_fcm$beta), names(lss_test_fcm$frequency))
     expect_true(is.null(lss_test_fcm$data))
     expect_identical(lss_test_fcm$terms, feat_test)
-    expect_identical(names(lss_test_fcm$seeds_weighted), names(seedwords("pos-neg")))
+    expect_identical(names(lss_test_fcm$seeds), names(seedwords("pos-neg")))
 
     expect_equal(
         names(lss_test_nd),
@@ -80,6 +80,25 @@ test_that("summary.textmodel_lss is working", {
 
     expect_silent(summary(lss_test))
     expect_silent(summary(lss_test_nd))
+
+})
+
+test_that("textmodel_lss print progress", {
+    skip_on_cran()
+
+    expect_output(
+        textmodel_lss(dfmt_test, seed, verbose = TRUE),
+        paste0("Calculating term-term similarity to 12 seed words...\n",
+               "Performing SVD by RSpectra...", collapse = ""),
+        fixed = TRUE
+    )
+
+    expect_output(
+        textmodel_lss(fcmt_test, seed, verbose = TRUE),
+        paste0("Calculating term-term similarity to 12 seed words...\n",
+               "Fitting GloVe model by rsparse...", collapse = ""),
+        fixed = TRUE
+    )
 
 })
 
@@ -232,11 +251,11 @@ test_that("textmodel_lss works with glob patterns", {
     dfmt <- dfm(toks_test)
     seed <- c("nice*" = 1, "positive*" = 1, "bad*" = -1, "negative*" = -1)
     lss <- textmodel_lss(dfmt, seed, k = 10)
-    expect_equal(names(lss$seeds_weighted), names(seed))
-    expect_equal(lengths(lss$seeds_weighted),
-                 c("nice*" = 0, "positive*" = 2, "bad*" = 3, "negative*" = 1))
-
-
+    expect_equal(names(lss$seeds), names(seed))
+    expect_equal(lss$seeds_weighted,
+                 c("positive" = 0.25, "positively" = 0.25,
+                   "badge" = -0.16, "bad" = -0.16, "badly" = -0.16,
+                   "negative" = -0.5), tolerance = 0.01)
 })
 
 test_that("textmodel_lss works with non-existent seeds", {
@@ -328,7 +347,6 @@ test_that("test smooth_lss", {
 
 test_that("works with single seed", {
     expect_silent(cohesion(lss_test))
-    expect_silent(strength(lss_test))
 })
 
 test_that("weight_seeds() works", {
@@ -365,3 +383,32 @@ test_that("weight_seeds() works", {
              "bb" = c("bb" = -1)),
     )
 })
+
+test_that("auto_weight is working", {
+
+    skip_on_cran()
+    lss1 <- textmodel_lss(dfmt_test, seed, k = 300)
+    lss2 <- textmodel_lss(dfmt_test, seed, k = 300, auto_weight = TRUE)
+    expect_true(
+        all(lss1$seeds_weighted != lss2$seeds_weighted)
+    )
+    expect_true(
+        all(sign(lss1$seeds_weighted) == sign(lss2$seeds_weighted))
+    )
+    expect_true(
+        all(abs(lss2$beta[names(lss2$seeds_weighted)] - lss1$seeds_weighted) < 0.05)
+    )
+
+    lss3 <- textmodel_lss(fcmt_test, seed, w = 50)
+    lss4 <- textmodel_lss(fcmt_test, seed, w = 50, auto_weight = TRUE)
+    expect_true(
+        all(lss3$seeds_weighted != lss4$seeds_weighted)
+    )
+    expect_true(
+        all(sign(lss3$seeds_weighted) == sign(lss4$seeds_weighted))
+    )
+    expect_true(
+        all(abs(lss4$beta[names(lss4$seeds_weighted)] - lss3$seeds_weighted) < 0.05)
+    )
+})
+