+++ tests/testthat/test-textplot.R
@@ -8,8 +8,7 @@ test_that("textplot_* works with SVD", {
     dfmt <- dfm(test_toks)
     seed <- c("nice*" = 1, "positive*" = 1, "bad*" = -1, "negative*" = -1)
     lss <- textmodel_lss(dfmt, seed, k = 10)
-    expect_equal(class(textplot_simil(lss, group = TRUE)), c("gg", "ggplot"))
-    expect_equal(class(textplot_simil(lss, group = FALSE)), c("gg", "ggplot"))
+    expect_equal(class(textplot_simil(lss)), c("gg", "ggplot"))
     expect_equal(class(textplot_terms(lss, highlighted = dict$keywords)),
                  c("gg", "ggplot"))
     expect_equal(class(textplot_terms(lss, highlighted = dict$keywords, max_words = 2)),
@@ -23,8 +22,7 @@ test_that("textplot_* works with Glove", {
     fcmt <- fcm(test_toks)
     seed <- c("nice*" = 1, "positive*" = 1, "bad*" = -1, "negative*" = -1)
     lss <- textmodel_lss(fcmt, seed, w = 10)
-    expect_equal(class(textplot_simil(lss, group = TRUE)), c("gg", "ggplot"))
-    expect_equal(class(textplot_simil(lss, group = FALSE)), c("gg", "ggplot"))
+    expect_equal(class(textplot_simil(lss)), c("gg", "ggplot"))
     expect_equal(class(textplot_terms(lss, highlighted = dict$keywords)),
                  c("gg", "ggplot"))
     expect_equal(class(textplot_terms(lss, highlighted = dict$keywords, max_words = 2)),
@@ -39,14 +37,6 @@ test_that("textplot_* raise error when attributes are missing", {
     coef <- rnorm(100)
     names(coef) <- topfeatures(dfmt, 100)
     lss <- as.textmodel_lss(coef)
-    expect_error(textplot_simil(lss, group = TRUE),
+    expect_error(textplot_simil(lss),
                  "textplot_simil() does not work with dummy models", fixed = TRUE)
-    expect_error(textplot_simil(lss, group = FALSE),
-                 "textplot_simil() does not work with dummy models", fixed = TRUE)
-})
-
-
-test_that("max_print is working", {
-
-
 })